Move code into grpc/io and driver module
