+++ src/grpc/driver/docker_container.rs
@@ -1,40 +1,28 @@
 use std::{
-    cmp,
-    collections::HashMap,
-    path::Path,
     pin::Pin,
-    sync::Arc,
     task::{Context, Poll},
     time::Duration,
 };
-//use std::io::Bytes;
 
-use bollard_buildkit_proto::moby;
 use bollard_stubs::models::{
-    ContainerWaitExitError, ContainerWaitResponse, ExecInspectResponse, HostConfig, Mount,
-    MountTypeEnum, SystemInfoCgroupDriverEnum, Volume,
+    ExecInspectResponse, HostConfig, Mount, MountTypeEnum, SystemInfoCgroupDriverEnum,
 };
-use bytes::{BufMut, Bytes, BytesMut};
-use futures_core::{ready, Future, TryStream};
+use bytes::BytesMut;
+use futures_core::Future;
 use futures_util::{StreamExt, TryStreamExt};
 use http::{
     header::{CONNECTION, UPGRADE},
     request::Builder,
     Method,
 };
-use hyper::Body;
-use pin_project_lite::pin_project;
-use tokio::io::{AsyncRead, AsyncWrite};
-use tokio_util::codec::FramedRead;
-use tonic::transport::server::Connected;
 use tower_service::Service;
 
 use crate::{
-    container::{Config, CreateContainerOptions, LogOutput, WaitContainerOptions},
+    container::{Config, CreateContainerOptions},
     errors::Error,
     exec::{CreateExecOptions, StartExecOptions, StartExecResults},
+    grpc::io::GrpcFramedTransport,
     image::CreateImageOptions,
-    read::NewlineLogOutputDecoder,
     Docker,
 };
 
@@ -42,162 +30,7 @@ const DEFAULT_IMAGE: &str = "moby/buildkit:master";
 const DEFAULT_STATE_DIR: &str = "/var/lib/buildkit";
 const DEFAULT_BUILDKIT_CONFIG_DIR: &str = "/etc/buildkit";
 
-pin_project! {
-    /// Reader for the [`into_async_read`](super::TryStreamExt::into_async_read) method.
-    #[derive(Debug)]
-    pub struct IntoAsyncRead<St>
-    where
-        St: TryStream<Error = std::io::Error>,
-        St::Ok: AsRef<[u8]>,
-    {
-        #[pin]
-        stream: St,
-        state: ReadState<St::Ok>,
-    }
-}
-
-#[derive(Debug)]
-enum ReadState<T: AsRef<[u8]>> {
-    Ready { chunk: T, chunk_start: usize },
-    PendingChunk,
-    Eof,
-}
-
-impl<St> IntoAsyncRead<St>
-where
-    St: TryStream<Error = std::io::Error>,
-    St::Ok: AsRef<[u8]>,
-{
-    pub(super) fn new(stream: St) -> Self {
-        Self {
-            stream,
-            state: ReadState::PendingChunk,
-        }
-    }
-}
-
-impl<St> AsyncRead for IntoAsyncRead<St>
-where
-    St: TryStream<Error = std::io::Error>,
-    St::Ok: AsRef<[u8]>,
-{
-    fn poll_read(
-        self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &mut tokio::io::ReadBuf<'_>,
-    ) -> Poll<std::io::Result<()>> {
-        let mut this = self.project();
-
-        loop {
-            match this.state {
-                ReadState::Ready { chunk, chunk_start } => {
-                    let chunk = chunk.as_ref();
-                    let len = cmp::min(buf.remaining(), chunk.len() - *chunk_start);
-
-                    buf.put_slice(&chunk[*chunk_start..*chunk_start + len]);
-
-                    *chunk_start += len;
-
-                    if chunk.len() == *chunk_start {
-                        *this.state = ReadState::PendingChunk;
-                    }
-
-                    return Poll::Ready(Ok(()));
-                }
-                ReadState::PendingChunk => match ready!(this.stream.as_mut().try_poll_next(cx)) {
-                    Some(Ok(chunk)) => {
-                        if !chunk.as_ref().is_empty() {
-                            *this.state = ReadState::Ready {
-                                chunk,
-                                chunk_start: 0,
-                            };
-                        }
-                    }
-                    Some(Err(err)) => {
-                        *this.state = ReadState::Eof;
-                        return Poll::Ready(Err(err));
-                    }
-                    None => {
-                        *this.state = ReadState::Eof;
-                        return Poll::Ready(Ok(()));
-                    }
-                },
-                ReadState::Eof => {
-                    return Poll::Ready(Ok(()));
-                }
-            }
-        }
-    }
-}
-
-#[allow(missing_debug_implementations)]
-/// TODO
-pub struct GrpcFramedTransport {
-    read: IntoAsyncRead<FramedRead<Pin<Box<dyn AsyncRead + Send>>, NewlineLogOutputDecoder>>,
-    write: Pin<Box<dyn AsyncWrite + Send>>,
-}
-
-impl AsRef<[u8]> for LogOutput {
-    fn as_ref(&self) -> &[u8] {
-        match self {
-            LogOutput::StdErr { message } => message.as_ref(),
-            LogOutput::StdOut { message } => message.as_ref(),
-            LogOutput::StdIn { message } => message.as_ref(),
-            LogOutput::Console { message } => message.as_ref(),
-        }
-    }
-}
-
-impl GrpcFramedTransport {
-    pub(crate) fn new(
-        read: Pin<Box<dyn AsyncRead + Send>>,
-        write: Pin<Box<dyn AsyncWrite + Send>>,
-        capacity: usize,
-    ) -> Self {
-        let output = FramedRead::with_capacity(read, NewlineLogOutputDecoder::new(true), capacity);
-        let read = IntoAsyncRead::new(output);
-        Self { read, write }
-    }
-}
-
-impl Connected for GrpcFramedTransport {
-    type ConnectInfo = ();
-
-    fn connect_info(&self) -> Self::ConnectInfo {}
-}
-
-impl AsyncRead for GrpcFramedTransport {
-    fn poll_read(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &mut tokio::io::ReadBuf<'_>,
-    ) -> Poll<std::io::Result<()>> {
-        Pin::new(&mut self.read).poll_read(cx, buf)
-    }
-}
-
-impl AsyncWrite for GrpcFramedTransport {
-    fn poll_write(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-        buf: &[u8],
-    ) -> Poll<std::io::Result<usize>> {
-        Pin::new(&mut self.write).poll_write(cx, buf)
-    }
-
-    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<std::io::Result<()>> {
-        Pin::new(&mut self.write).poll_flush(cx)
-    }
-
-    fn poll_shutdown(
-        mut self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-    ) -> Poll<Result<(), std::io::Error>> {
-        Pin::new(&mut self.write).poll_shutdown(cx)
-    }
-}
-
-impl Service<http::Uri> for Driver {
+impl Service<http::Uri> for DockerContainer {
     type Response = GrpcFramedTransport;
     type Error = Error;
     type Future = Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + Send>>;
@@ -209,7 +42,6 @@ impl Service<http::Uri> for Driver {
     fn call(&mut self, _req: http::Uri) -> Self::Future {
         let client = Docker::clone(&self.docker);
         let name = String::clone(&self.name);
-        let session_id = String::clone(&self.session_id);
 
         let fut = async move {
             let exec_id = client
@@ -252,22 +84,21 @@ impl Service<http::Uri> for Driver {
                 })
         };
 
-        // Return the response as an immediate future
         Box::pin(fut)
     }
 }
 
 /// TODO
 #[derive(Debug)]
-pub struct DriverBuilder {
-    inner: Driver,
+pub struct DockerContainerBuilder {
+    inner: DockerContainer,
 }
 
-impl DriverBuilder {
+impl DockerContainerBuilder {
     /// TODO
     pub fn new(name: &str, docker: &Docker, session_id: &str) -> Self {
         Self {
-            inner: Driver {
+            inner: DockerContainer {
                 name: String::from(name),
                 docker: Docker::clone(docker),
                 session_id: String::from(session_id),
@@ -281,7 +112,7 @@ impl DriverBuilder {
     }
 
     /// TODO
-    pub async fn bootstrap(mut self) -> Result<Driver, Error> {
+    pub async fn bootstrap(mut self) -> Result<DockerContainer, Error> {
         debug!("booting buildkit");
 
         if self.inner.net_mode.is_none() {
@@ -311,7 +142,7 @@ impl DriverBuilder {
     }
 
     /// TODO
-    pub fn network(&mut self, net: &str) -> &mut DriverBuilder {
+    pub fn network(&mut self, net: &str) -> &mut DockerContainerBuilder {
         if net == "host" {
             self.inner
                 .args
@@ -321,11 +152,36 @@ impl DriverBuilder {
         self.inner.net_mode = Some(net.to_string());
         self
     }
+
+    /// The image to use when spinning up a `Buildkit` container. The default is [`DEFAULT_IMAGE`]
+    pub fn image(&mut self, image: &str) -> &mut DockerContainerBuilder {
+        self.inner.image = Some(String::from(image));
+        self
+    }
+
+    /// The cgroup to attach to - by default all `Buildkit` containers are placed under the same
+    /// cgroup so that limits are applied across the whole host
+    pub fn cgroup_parent(&mut self, cgroup_parent: &str) -> &mut DockerContainerBuilder {
+        self.inner.cgroup_parent = Some(String::from(cgroup_parent));
+        self
+    }
+
+    /// Set an env variable for the `Buildkit` container.
+    pub fn env(&mut self, env: &str) -> &mut DockerContainerBuilder {
+        self.inner.env.push(String::from(env));
+        self
+    }
+
+    /// Set a additional run command arguments to the `Buildkit` docker execution.
+    pub fn arg(&mut self, arg: &str) -> &mut DockerContainerBuilder {
+        self.inner.args.push(String::from(arg));
+        self
+    }
 }
 
 /// TODO
 #[derive(Debug)]
-pub struct Driver {
+pub struct DockerContainer {
     /// TODO
     name: String,
     /// TODO
@@ -344,7 +200,7 @@ pub struct Driver {
     args: Vec<String>,
 }
 
-impl Driver {
+impl DockerContainer {
     /// TODO
     pub async fn create(&self) -> Result<(), Error> {
         let image_name = if let Some(image) = &self.image {
